+++ pilz_trajectory_generation/include/pilz_trajectory_generation/command_list_manager.h
@@ -20,7 +20,7 @@
 #include <moveit/planning_interface/planning_interface.h>
 #include <moveit_msgs/MotionPlanResponse.h>
 
-#include "pilz_msgs/MotionBlendRequestList.h"
+#include "pilz_msgs/MotionSequenceRequest.h"
 #include "pilz_trajectory_generation/trajectory_blender.h"
 
 namespace pilz_trajectory_generation {
@@ -29,7 +29,7 @@ namespace pilz_trajectory_generation {
  * @brief The CommandListManager class
  * This class can create a smooth trajectory from a given list of motion commands.
  * The trajectory generated from the motion commands are blended with each other within a blend radius given
- * within the MotionBlendRequestList.
+ * within the MotionSequenceRequest.
  */
 class CommandListManager {
 
@@ -49,7 +49,7 @@ public:
    * @return True if the generation was successful, false otherwise
    */
   bool solve(const planning_scene::PlanningSceneConstPtr& planning_scene,
-             const pilz_msgs::MotionBlendRequestList& req_list,
+             const pilz_msgs::MotionSequenceRequest& req_list,
              planning_interface::MotionPlanResponse &res);
 
 private:
@@ -63,7 +63,7 @@ private:
    * @param res The response used to set the error code on validation error
    * @return True if all conditions hold, false otherwise
    */
-  bool validateRequestList(const pilz_msgs::MotionBlendRequestList &req_list, planning_interface::MotionPlanResponse& res);
+  bool validateRequestList(const pilz_msgs::MotionSequenceRequest &req_list, planning_interface::MotionPlanResponse& res);
 
   /**
    * @brief Validates that two consecutive blending radii do not overlap
@@ -87,7 +87,7 @@ private:
    * @return True if trajectories for all request could be generated
    */
   bool solveRequests(const planning_scene::PlanningSceneConstPtr& planning_scene,
-                     const pilz_msgs::MotionBlendRequestList &req_list,
+                     const pilz_msgs::MotionSequenceRequest &req_list,
                      planning_interface::MotionPlanResponse &res,
                      std::vector<planning_interface::MotionPlanResponse>& motion_plan_responses,
                      std::vector<double>& radii);